move group operation profiling code into its own executable
